tooling: Move confirmation_key view to development only views.
(Previously) zerver.views.registration.confirmation_key was relevant
to development urls only; It is moved to dev-views in this commit

Fixes #11256
